Fixed integer cast in go.builtin.fatal for exit code

+++ src/gocore/golib/builtin.cs
@@ -445,7 +445,7 @@ namespace go
             throw new InvalidOperationException($"{message} [{code}]");
         #else
             Console.Error.WriteLine(message);
-            Environment.Exit(code);
+            Environment.Exit((int)code);
         #endif
         }
